backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via redis
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.2
    # via -r requirements.in
google-auth==2.35.0
    # via google-auth-oauthlib
google-auth-oauthlib==1.2.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
hiredis==3.0.0
    # via pydantic-redis
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   requests
oauthlib==3.2.2
    # via requests-oauthlib
orjson==3.10.9
    # via pydantic-redis
packaging==24.1
    # via gunicorn
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic[email]==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-redis
pydantic-core==2.23.4
    # via pydantic
pydantic-redis==0.7.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
redis==5.1.1
    # via pydantic-redis
requests==2.32.3
    # via requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.40.0
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.32.0
    # via -r requirements.in
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
